Beginning of the nearpi algorithm

+++ tools/cont_frac.hpp
@@ -17,12 +17,12 @@
 class ContFrac
 {
 public:
-    typedef unsigned long coeff_type;
+    typedef __uint128_t coeff_type;
     static const coeff_type s_inf = (coeff_type)-1;
     
     bool load_str(const char *str);
     bool load_coeffs(const std::vector<coeff_type> coeffs);
-    coeff_type get_coeff(int ndx) { return ndx < m_coeff.size() ? m_coeff[ndx] : s_inf; }
+    coeff_type get_coeff(size_t ndx) { return ndx < m_coeff.size() ? m_coeff[ndx] : s_inf; }
     bool apply_matrix(long int a, long int b, long int c, long int d);
     bool invert();
     size_t size() { return m_coeff.size(); }
